backend(search functionality): Connected the backend search with the frontend
