+++ onemed1a-frontend/src/api/mediaClient.js
@@ -2,6 +2,6 @@ import apiClient from "./apiClient/apiClient.js";
 
 
 export async function getMediaById(id) {
-    const response = await apiClient.get(`media/${id}`);
+    const response = await apiClient.get(`/media/${id}`);
     return response.data;
 }
