+++ onemed1a-frontend/src/api/searchAPI.js
@@ -1,0 +1,9 @@
+import apiClient from "./apiClient.js";
+
+
+export async function suggest(text) {
+    const response = await apiClient.get(`/media/suggest`,  {
+        params: { q: text }
+    });
+    return response.data;
+}
